Make OAS_PATH default to the current directory and remove appended path (#2)
* Make OAS_PATH default to the current directory and remove appended path

Previously we added _open_api/api_specs to OAS_PATH, but this makes the
gem difficult to use locally. This PR removes this requirement and Nexmo
Developer will set an OAS_PATH of _open_api/api_specs.

This allows users to run `nexmo-oas-renderer` locally in the
api-specification repo and have the endpoint rendered as expected. This
does not include any errors as they are provided in `errors.md` in the
Nexmo Developer project, and will be addressed in a future PR

* Markdown filters don't need to look in OAS_PATH