Fix allOf in RequestBody entries
We previously handled allOf in request bodies manually
as the OAS Parser did not support allOf. The parser has
since added support which meant that we could end up trying
to call handle_all_of on an array, rather than a hash containing
an allOf entry

The solution for this is to remove the hack previously added. It
has been tested on all existing specs and they do not throw a 500